Add files via upload
The purpose of the code is to demonstrate two different approaches for performing Mean Curvature Flow (MCF) on a 3D mesh. Mean Curvature Flow is a geometric process where the vertices of the mesh are iteratively adjusted to smooth the surface. This is done by moving each vertex in the direction of the mean curvature normal vector, effectively reducing surface irregularities over time. I used the gp toolbox. The approaches used are: Explicit Euler and the Semi-Implicit Desbrun's Method.